WIP converting to float32 and index based

+++ src/main/java/ai/iambot/elasticsearch/script/metrics/ChiSquaredMetric.java
@@ -1,15 +1,15 @@
 package ai.iambot.elasticsearch.script.metrics;
 
 public class ChiSquaredMetric implements Metric{
-    double[] inputVector;
-    double epsilon = 1e-10;
-    public ChiSquaredMetric(double[] inputVector){
+    float[] inputVector;
+    float epsilon = 1e-10;
+    public ChiSquaredMetric(float[] inputVector){
         this.inputVector = inputVector;
     }
 
     @Override
-    public double metric(double[] queryVector) {
-        double result = 0.0;
+    public float metric(float[] queryVector) {
+        float result = 0.0;
         for (int i = 0; i < queryVector.length; i++) {
             double diff = queryVector[i]-this.inputVector[i];
             result += (diff * diff)/(queryVector[i]+inputVector[i]+epsilon);
