+++ src/main/java/ai/iambot/elasticsearch/script/metrics/ChiSquaredMetricProvider.java
@@ -2,7 +2,7 @@ package ai.iambot.elasticsearch.script.metrics;
 
 public class ChiSquaredMetricProvider implements MetricProvider{
     @Override
-    public Metric getMetric(double[] inputVector) {
+    public Metric getMetric(float[] inputVector) {
         return new ChiSquaredMetric(inputVector);
     }
 }
